+++ js/login.js
@@ -16,7 +16,6 @@ const remember = document.getElementById("remember");
 const btnLabel = document.getElementById("btnLabel");
 const btnSpinner = document.getElementById("btnSpinner");
 const eyeToggle = document.getElementById("eyeToggle");
-
 // Check if email is valid by using regex
 function isEmailValid(email) {
   const pattern = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
@@ -60,7 +59,7 @@ function hideAllErrors() {
 // Returns true if everything is valid
 function validate(email, password) {
   let valid = true;
-
+  
   if (!email) {
     showFieldError("email", "Please enter your email address.");
     valid = false;
@@ -68,7 +67,7 @@ function validate(email, password) {
     showFieldError("email", "Please enter a valid email address.");
     valid = false;
   }
-
+  
   if (!password) {
     showFieldError("password", "Please enter your password.");
     valid = false;
@@ -76,7 +75,7 @@ function validate(email, password) {
     showFieldError("password", "Password must be at least 6 characters.");
     valid = false;
   }
-
+  
   return valid;
 }
 
@@ -84,6 +83,7 @@ function setLoading(isLoading) {
   continueBtn.disabled = isLoading;
   continueBtn.classList.toggle("loading", isLoading);
 }
+module.exports = {setLoading};
 
 eyeToggle.addEventListener("click", () => {
   const isHidden = passInput.type === "password";
