+++ package.json
@@ -19,10 +19,17 @@
     "test:watch": "jest --watchAll"
   },
   "jest": {
-    "testEnvironment": "jsdom"
+    "reporters": [
+      "default",
+      ["jest-html-reporter", {
+        "pageTitle": "Test Report",
+        "outputPath": "./test-report.html"
+      }]
+    ]
   },
   "devDependencies": {
     "jest": "^30.2.0",
-    "jest-environment-jsdom": "^30.2.0"
+    "jest-environment-jsdom": "^30.2.0",
+    "jest-html-reporter": "^4.3.0"
   }
 }
